feat: Update TPA after updating a catalog
