+++ src/pages/catalog/NewCatalog.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { generateTPA } from './TpaUtils';
 import { Form, Card, Row, Col, Button } from "react-bootstrap";
 import { useNavigate } from "react-router-dom";
 import { useCatalogs } from "../../hooks/useCatalogs";
@@ -10,7 +11,6 @@ import {
 } from "../../features/controls/controlSlice";
 import { removeInput } from "../../features/inputs/inputSlice";
 import ControlForm from "./ControlForm";
-import tpaTemplate from "./tpa-template.json";
 import { useMashups } from "../../hooks/useMashups";
 
 function NewCatalog() {
@@ -27,9 +27,9 @@ function NewCatalog() {
   const dispatch = useDispatch();
   const controls = useSelector((state) => state.controls.controls);
   const navigate = useNavigate();
+  const { getMashupByIdFromTheDB } = useMashups();
 
   const { createControlInDB, createControlInputInDB } = useControls();
-  const { getMashupByIdFromTheDB } = useMashups();
 
   // Handler function for remove a control
   const handleRemoveControl = (controlId) => {
@@ -88,7 +88,7 @@ function NewCatalog() {
       );
       await Promise.all(controlPromises);
 
-      generateTPA(controls, catalogId);
+      generateTPA(controls, catalogId, getMashupByIdFromTheDB, inputs);
 
       navigate("/catalogs");
     } catch (error) {
@@ -100,138 +100,7 @@ function NewCatalog() {
     dispatch(addEmptyControl());
   };
 
-  const toUpperSnakeCase = (name) => {
-    return name.toUpperCase().replace(/\s+/g, "_");
-  };
-
-  const generateTPA = async (controls, catalogId) => {
-    const tpa = JSON.parse(JSON.stringify(tpaTemplate));
-
-    for (const control of controls) {
-      const countingMetricName = "NUMBER_OF_ELEMENTS_FOR_" + toUpperSnakeCase(control.name);
-      const positiveElementsMetricName = "NUMBER_OF_POSITIVE_ELEMENTS_FOR_" + toUpperSnakeCase(control.name);
-      const guaranteeName = "RATE_FOR_" + toUpperSnakeCase(control.name);
-      const mashup = await getMashupByIdFromTheDB(control.mashup_id);
-
-      let config = {};
-      for (const input of inputs.inputs[control.id]) {
-        config[input.name] = input.value;
-      }
-      let trueConfig = { ...config };
-      trueConfig["Status"] = "true";
-
-      addMetricToTPA(tpa, countingMetricName, mashup, config);
-      addMetricToTPA(tpa, positiveElementsMetricName, mashup, trueConfig);
-      addGuaranteeToTPA(tpa, guaranteeName, control, countingMetricName, positiveElementsMetricName, 50);
-    }
-
-    // Convierte el objeto TPA a una cadena JSON
-    const tpaString = JSON.stringify(tpa, null, 2);
-
-    saveTpa(tpaString, catalogId);
-  };
-
-  const addMetricToTPA = async (tpa, metricName, mashup, config) => {
-    tpa.terms.metrics[metricName] = {
-      collector: {
-        infrastructurePath: "internal.collector.events",
-        endpoint: "/api/v2/computations",
-        type: "POST-GET-V1",
-        config: {
-          scopeManager:
-            "http://host.docker.internal:5700/api/v1/scopes/development",
-        },
-      },
-      measure: {
-        computing: "actual",
-        element: "number",
-        event: {
-          status: {
-            [mashup.name]: {},
-          },
-        },
-        config: config,
-        scope: {
-          project: {
-            name: "Project",
-            description: "Project",
-            type: "string",
-            default: "example-project",
-          },
-          class: {
-            name: "Class",
-            description: "Group some projects",
-            type: "string",
-            default: "example-class",
-          },
-        },
-      },
-    };
-  }
-
-  const addGuaranteeToTPA = async (tpa, guaranteeName, control, countingMetricName, positiveElementsMetricName, objective) => {
-    tpa.terms.guarantees.push({
-      id: guaranteeName,
-      notes: `#### Description\n\`\`\`\n${control.description}`,
-      description: control.description,
-      scope: {
-        project: {
-          name: "Project",
-          description: "Project",
-          type: "string",
-          default: "example-project",
-        },
-        class: {
-          name: "Class",
-          description: "Group some projects",
-          type: "string",
-          default: "example-class",
-        },
-      },
-      of: [
-        {
-          scope: {
-            project: "example-project",
-          },
-          objective: `${positiveElementsMetricName}/${countingMetricName} * 100 > ${objective}`,
-          with: {
-            [countingMetricName]: {},
-            [positiveElementsMetricName]: {},
-          },
-          window: {
-            type: "static",
-            period: control.period.toLowerCase(),
-            initial: control.startDate,
-            end: control.endDate,
-          },
-        },
-      ],
-    });
-  }
-
-  const saveTpa = async (tpaContent, catalogId) => {
-    try {
-      const response = await fetch("http://localhost:3001/api/catalogs/tpa", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          content: tpaContent,
-          catalog_id: catalogId,
-        }),
-      });
-
-      if (response.ok) {
-        console.log("TPA guardado en el servidor");
-      } else {
-        console.error("Error al guardar el TPA:", response.statusText);
-      }
-    } catch (error) {
-      console.error("Error al realizar la solicitud:", error);
-    }
-  };
-
+  
   // JSX representing the component's UI
   return (
     <div className="container">
